Add ArrayCombineFunctionThrowTypeExtension
