+++ tests/PHPStan/Rules/Exceptions/ThrowsVoidMethodWithExplicitThrowPointRuleTest.php
@@ -8,6 +8,7 @@ use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\RequiresPhp;
 use ThrowsVoidMethod\MyException;
 use UnhandledMatchError;
+use ValueError;
 
 /**
  * @extends RuleTestCase<ThrowsVoidMethodWithExplicitThrowPointRule>
@@ -97,6 +98,13 @@ class ThrowsVoidMethodWithExplicitThrowPointRuleTest extends RuleTestCase
 		$this->missingCheckedExceptionInThrows = $missingCheckedExceptionInThrows;
 		$this->checkedExceptionClasses = $checkedExceptionClasses;
 		$this->analyse([__DIR__ . '/data/throws-void-method.php'], $errors);
+	}
+
+	public function testBug13642(): void
+	{
+		$this->missingCheckedExceptionInThrows = false;
+		$this->checkedExceptionClasses = [ValueError::class];
+		$this->analyse([__DIR__ . '/data/bug-13642.php'], []);
 	}
 
 	#[RequiresPhp('>= 8.0')]
